requirements.txt
sqlalchemy==2.0.32
pandas==2.2.2
pyodbc==5.1.0

pydantic==2.8.2
pydantic-settings==2.4.0

sqlcipher3-wheels